fix: TAREA-10 -Werror activo — 0 errors 0 warnings en todo el pipeline

+++ ml-detector/src/contract_validator.cpp
@@ -86,7 +86,7 @@ void ContractValidator::log_missing_features(const protobuf::NetworkSecurityEven
         const google::protobuf::Reflection* reflection = ddos.GetReflection();
         std::vector<const google::protobuf::FieldDescriptor*> fields;
         reflection->ListFields(ddos, &fields);
-        ddos_fields = fields.size();
+        ddos_fields = static_cast<int>(fields.size());
         
         if (ddos_fields < 10) {
             logger->warn("  ddos_embedded incomplete: {} / 10 fields", ddos_fields);
@@ -102,7 +102,7 @@ void ContractValidator::log_missing_features(const protobuf::NetworkSecurityEven
         const google::protobuf::Reflection* reflection = ransomware.GetReflection();
         std::vector<const google::protobuf::FieldDescriptor*> fields;
         reflection->ListFields(ransomware, &fields);
-        ransomware_fields = fields.size();
+        ransomware_fields = static_cast<int>(fields.size());
         
         if (ransomware_fields < 10) {
             logger->warn("  ransomware_embedded incomplete: {} / 10 fields", ransomware_fields);
@@ -118,7 +118,7 @@ void ContractValidator::log_missing_features(const protobuf::NetworkSecurityEven
         const google::protobuf::Reflection* reflection = traffic.GetReflection();
         std::vector<const google::protobuf::FieldDescriptor*> fields;
         reflection->ListFields(traffic, &fields);
-        traffic_fields = fields.size();
+        traffic_fields = static_cast<int>(fields.size());
         
         if (traffic_fields < 10) {
             logger->warn("  traffic_classification incomplete: {} / 10 fields", traffic_fields);
@@ -134,7 +134,7 @@ void ContractValidator::log_missing_features(const protobuf::NetworkSecurityEven
         const google::protobuf::Reflection* reflection = internal.GetReflection();
         std::vector<const google::protobuf::FieldDescriptor*> fields;
         reflection->ListFields(internal, &fields);
-        internal_fields = fields.size();
+        internal_fields = static_cast<int>(fields.size());
         
         if (internal_fields < 10) {
             logger->warn("  internal_anomaly incomplete: {} / 10 fields", internal_fields);
